Remove editable attribute of plotly js plot
Editing labels and titles directly on a plotly plot is not a function of react-chart-editor but one native to plotly.js. As a result changes made to the plot in this way do not trigger a change in the original plot (this is expected since react-chart-editor has no idea these changes have been made). Until a solution is found it's better to disable this to avoid confusion for users.